pyproject.toml
[project]
name = "llm-memory"
version = "0.1.0"
description = "Persistent shared memory system for multi-agent AI"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastmcp>=2.0.0",
    "lancedb>=0.5.0",
    "pyarrow>=14.0.0",
    "sentence-transformers>=2.2.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "filelock>=3.12.0",
    "pydantic>=2.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
llm-memory = "src.mcp_server.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
